server/requirements.txt
pdf2docx>=0.5.12
python-docx>=1.2.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
PyMuPDF>=1.22.0
